+++ hggit/compat.py
@@ -1,5 +1,22 @@
 from dulwich.object_store import PackBasedObjectStore
 
+try:
+    # Mercurial >= 7.2
+    from mercurial.bundle2_part_handlers import parthandler
+    from mercurial.interfaces.repository import IPeer
+except ImportError:
+    # fallback for Mercurial < 7.2
+    from mercurial.bundle2 import parthandler
+    from mercurial.interfaces.repository import peer as IPeer
+
+    # the following is to silence pyflakes, which has no support for
+    # "noqa: F401".
+    #
+    # TODO: when possible (for example once we migrate to ruff), replace the
+    #       following asserts with a "noqa: F401" directive in the imports above
+    assert parthandler
+    assert IPeer
+
 
 if hasattr(PackBasedObjectStore, 'delete_loose_object'):