+++ NEWS.rst
@@ -6,7 +6,7 @@ This is a feature release, although without the actual features.
 * Fix an error on pull with commits missing on Mercurial side. We'd
   try to advance phase boundaries for them, which failed.
 * Drop support for Python 3.9 & Dulwich 0.21
-* Add support for Dulwich 0.24 and Mercurial 7.1.
+* Add support for Dulwich 0.24 and Mercurial 7.2.
 * Suppress pointless warning when pushing deleted tags. (#446)
 
 This release requires Mercurial 6.6, or later, Dulwich 0.22.1 or later